#9171 move LitTemplateParserImpl into the internal subpackage
