+++ flow-lit-template/src/main/java/com/vaadin/flow/component/littemplate/LitTemplateParser.java
@@ -15,6 +15,7 @@
  */
 package com.vaadin.flow.component.littemplate;
 
+import com.vaadin.flow.component.littemplate.internal.LitTemplateParserImpl;
 import org.jsoup.nodes.Element;
 
 import com.vaadin.flow.di.Instantiator;
